// src/features/match/types/match.types.ts
export interface Match {
  idEvent: string;
  idAPIfootball?: string | null;

  strEvent: string;
  strEventAlternate?: string | null;
  strFilename?: string | null;
  strSport?: string | null;

  idLeague?: string | null;
  strLeague?: string | null;
  strLeagueBadge?: string | null;

  strSeason?: string | null;
  strDescriptionEN?: string | null;

  strHomeTeam: string;
  strAwayTeam: string;

  intHomeScore?: string | null;
  intAwayScore?: string | null;
  intRound?: string | null;

  intSpectators?: string | null;
  strOfficial?: string | null;

  strTimestamp?: string | null;
  dateEvent?: string | null;
  dateEventLocal?: string | null;
  strTime?: string | null;
  strTimeLocal?: string | null;

  strGroup?: string | null;

  idHomeTeam: string;
  strHomeTeamBadge?: string | null;

  idAwayTeam: string;
  strAwayTeamBadge?: string | null;

  intScore?: string | null;
  intScoreVotes?: string | null;

  strResult?: string | null;

  idVenue?: string | null;
  strVenue?: string | null;

  strCountry?: string | null;
  strCity?: string | null;

  strPoster?: string | null;
  strSquare?: string | null;

  strFanart?: string | null;
  strThumb?: string | null;
  strBanner?: string | null;

  strMap?: string | null;

  strTweet1?: string | null;
  strTweet2?: string | null;
  strTweet3?: string | null;
  strVideo?: string | null;

  strStatus?: string | null;
  strPostponed?: string | null;

  strLocked?: string | null;
}

export interface LiveMatchesResponse {
  events: Match[] | null;
}